iommu: Validate that pci_for_each_dma_alias() matches the groups
Directly check that the devices returned by pci_for_each_dma_alias() are
in the same group, or not yet matched to any group. This helps validate
that pci_for_each_dma_alias() and pci_bus_isolated() are consistent.

This should eventually be hidden behind a debug kconfig, but for now it is
good to get feedback from more diverse systems if there are any problems.

Signed-off-by: Jason Gunthorpe <jgg@nvidia.com>