iommu/tegra: Remove manual of parsing
This driver calls iommu_device_register() with a non-NULL so the only
way the core code will call probe_device is if the iommus DT property
was successfully parsed.

None of the DTS have more than one iommu_device per iommus property, so
this is a normal driver.

For some reason the driver manually parses it again. Remove all of this
and just use the normal of_xlate and probe_device_fwspec() to get
the instance.

The driver was also using the of_xlate to set the iommu_device in the
dev_iommu_priv.

Signed-off-by: Jason Gunthorpe <jgg@nvidia.com>