cover-letter: Simplify device probing for fwspec drivers
These days the core code has already computed the iommu_device for the
fwspec that the driver will be bound to. In many simple cases there is
only one iommu_driver per fwspec so it can be passed directly from the
core code to the driver in a new probe_device_fwspec() function.

This frees the driver from having to figure out what the iommu_device
instance is for the device.

Convert most of the drivers to use this mechanism. The following are left
out:

 apple-dart
 exynos
 msm
 omap
   These are all multi-instance drivers, meaning their iommus property
   specifies multiple iommu_device's

 fsl_pamu
 amd
 intel
 s390
 xen
   Not fwspec drivers

 iommufd
   Will be changed after the merge window

There are several themes in the conversions:
 smmu
 smmu-v3
 riscv
 virtio
   Have some code to convert the fwpsec back to the iommu_driver open
   coded. eg using bus_find_device_by_fwnode. Just use container_of()
   on the argument to probe_device_fwspec().

 mtk
 vmsa
 sprd
 sun50i
 qcom
   Using something like of_find_device_by_node() in the of_xlate()
   to find the iommu_driver and then stored it in the dev_iommu_priv for
   probe_device to pick up. Just use container_of()
   on the argument to probe_device_fwspec().

 rockchip
   Like the above except the also allocate the per-device struct in
   of_xlate() too. This will leak memory in various error cases. Move
   the memory allocation to probe and use container_of()
   on the argument to probe_device_fwspec().

 tegra
 mtk_v1
   For some reason this repeats all the parsing. Use of_xlate to build a
   fwspec and operate like a normal driver. None of these seem to have
   multi-iommu_device DTs in the tree.

Cc: Johan Hovold <johan@kernel.org>
Signed-off-by: Jason Gunthorpe <jgg@nvidia.com>