cover-letter: Add alloc_container()
Consolidate several copies of this similar pattern into centralized macros
with some documentation.

alloc_container() provides a helper for subsystems to implement an "alloc
and init" pattern where the subsystem struct and the driver struct have a
container_of relationship:

  struct subsystem {...};
  struct driver {struct subsystem subsys; ...}

  struct driver *drv;
  drv = alloc_container(struct driver, subsys, subsystem_alloc, ...);
  if (IS_ERR(drv))
       return PTR_ERR(drv);
  [..]
  rc = subsystem_register(&drv->subsys);
  if (rc) {
	subsystem_put(&drv->subsys);
	goto err;
  }

Convert several subsystems that already have this pattern open coded to
use the new macros.

This pattern is popular because of how consistent and safe it is for the
drivers to use.

Compared to the init type pattern:

  struct driver *drv = kzalloc(sizeof(struct driver), GFP_KERNEL);
  rc = subsystem_init(&drv->subsys);
  if (rc) {
	kfree(drv);
	goto err;
  }
  [..]
  rc = subsystem_register(&drv->subsys);
  if (rc) {
	subsystem_put(&drv->subsys);
	goto err;
  }

The driver is not exposed to an uninitialized struct subsystem, nor does
it have to figure out when to use kfree and when to use
subsystem_put(). Both of these are common sources of bugs in error paths
in drivers.

Combined with the approach of having init/deinit ops passed to the alloc
it puts the subsystem in charge of all the error unwinding and sequencing
scenarios rather than open coding this frequently tricky code in drivers.

Signed-off-by: Jason Gunthorpe <jgg@nvidia.com>