dma-buf: Add the SGT DMA mapping type
The SGT DMA mapping type represents the existing scatterlist based DMA
mapping. It is used to transparently wrap non-mapping type aware drivers
into the mapping type system.

A mapping type aware driver can include SGT in it's match list and the
core code will automatically match it to non-aware drivers and make it
work as though the other driver natively supported matching.

The SGT type handles the allow_peer2peer flag directly through matching
logic. The importer indicates if it is willing to accept peer2peer and
the exporter indicates if it requires peer2peer. A required peer2peer
exporter will not match to an importer that does not accept peer2peer.

Signed-off-by: Jason Gunthorpe <jgg@nvidia.com>