dma-buf: Add the Physical Address List DMA mapping type
This type is required by iommufd and kvm as dmabuf importers.

Due to sensitivity about abusing physical addresses, restrict importers by
using EXPORT_SYMBOL_FOR_MODULES(). Only iommufd can implement an importer,
the kernel module loader will enforce this.

Allow anything to implement an exporter as there are use cases in
DPDK/SPDK to connect GPU memory into VFIO/iommufd and it is hard to abuse
the API as an exporter.

The physical address list exporter returns a physical address list in a
simple kvalloc'd array of struct phys_vec.

For now all entries are assumed to be assumed to be MMIO and iommufd
will map into the IOMMU using the IOMMU_MMIO flag.

Signed-off-by: Jason Gunthorpe <jgg@nvidia.com>