cover-letter: Add dmabuf mapping types and convert vfio/iommufd to use them
DMA BUF mappings types allow the importer and exporter to negotiate the
format of the map/unmap during the attachment.

Currently DMA BUF only support struct scatterlist as the attachment
type, this is not sufficient for all use cases as dma_addr_t is a very
specific and limited type. Some drivers are already working around this by
hacking a detection if the same driver is importing/exporting and then
placing something driver-specific in the dma_addr_t.

With mapping types the importing driver can declare what it supports:

	struct dma_buf_mapping_match imp_match[] = {
		DMA_BUF_IMAPPING_MY_DRIVER(dev, ...),
		DMA_BUF_IMAPPING_SGT(dev, false),
	};
	attach = dma_buf_mapping_attach(dmabuf, imp_match, ...)

And the exporting driver can declare what it supports:

int exporter_match_mapping(struct dma_buf_match_args *args)
{
	struct dma_buf_mapping_match exp_match[] = {
		DMA_BUF_EMAPPING_MY_DRIVER(my_ops, dev, ...),
		DMA_BUF_EMAPPING_SGT(sgt_ops, dev, false),
	};
	return dma_buf_match_mapping(args, exp_match, ...);
}

During dma_buf_mapping_attach() the core code will select a mutal match
between the importer and exporter and record it in the attach->map_type

Each mapping type has its own types and function calls for
mapping/unmapping. As such they can provide an interface that is perfectly
tailored to their purpose.

To support this the exporter provides a set of ops for each mapping type
that replaces the attach/detach/map/unmap ops in dma_buf_ops. The core
code arranges to call the right ops based on the selected match.

Signed-off-by: Jason Gunthorpe <jgg@nvidia.com>